Refactor code, therefore reduce technical debt.
I added code that clears the memory buffer which the user command is
written to. Also added code that checks the length of prompts and user
commands before running them, instead of relying on statically inputted
values. Lastly, made sure to push all non-volatile registers to the
stack; not just the ones im using.